backend/requirements.txt
# Backend dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2

# Document processing
pypdf2==3.0.1
python-docx==1.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# Vector database
chromadb==0.4.18
faiss-cpu==1.7.4

# Embedding and AI
langchain==0.0.335
langchain-google-genai==0.0.5
google-generativeai==0.3.1
sentence-transformers==2.2.0
huggingface-hub==0.21.4
# OCR
pytesseract==0.3.10
easyocr==1.7.1

# Speech processing
speechrecognition==3.10.0
gTTS==2.3.2
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
numpy>=1.26.0
pandas==2.1.1
langdetect==1.0.9

# CORS
fastapi-cors==0.0.6